refactor: split window/mod.rs into focused submodules
Extract 63 methods from the monolithic impl Window block into 5
focused submodules:

  actions.rs   — keyboard shortcuts, focus, zoom, clipboard (347 lines)
  terminal.rs  — terminal lifecycle, split, close, recovery (689 lines)
  dialogs.rs   — confirmation dialogs, about, rename, bookmark (265 lines)
  sidebar.rs   — bookmark/command sidebar, activity, subtitles (378 lines)
  input.rs     — input sync, preferences, color scheme (111 lines)

mod.rs retains the imp module, GObject boilerplate, construction,
state load/save, setup_signals, and session lifecycle (963 lines).

Cross-file method calls use pub(super) visibility — no public API
changes to Window.

Fixes #98
